# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API documentation for LPM",
    "version": "1.0.5",
    "title": "LPM",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "artemstrigkov@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://ec2-35-183-44-100.ca-central-1.compute.amazonaws.com:8080"
    }
  },
  "host": "ec2-35-182-71-219.ca-central-1.compute.amazonaws.com:4000",
  "basePath": "/api/v1",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/locations": {
      "get": {
        "tags": [
          "LOCATIONS"
        ],
        "summary": "Locations listing",
        "description": "",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Invalid username supplied"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "post": {
        "tags": [
          "LOCATIONS"
        ],
        "summary": "Locations",
        "description": "Locations",
        "operationId": "Locations",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Locations",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Location"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Invalid username supplied"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "patch": {
        "tags": [
          "LOCATIONS"
        ],
        "summary": "Locations Update",
        "description": "Locations Update",
        "operationId": "Locations Update",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Locations",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LocationUpdate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Invalid username supplied"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "delete": {
        "tags": [
          "LOCATIONS"
        ],
        "summary": "Locations delete",
        "description": "Locations delete",
        "operationId": "Locations delete",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Locations",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Invalid username supplied"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    },
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "https://petstore.swagger.io/oauth/authorize",
      "flow": "implicit",
      "scopes": {
        "read:pets": "read your pets",
        "write:pets": "modify pets in your account"
      }
    }
  },
  "definitions": {
...
    "Id": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "1"
        }
      },
      "xml": {
        "name": "Id"
      }
    },
...
    "Location": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "CF Sherway gardens"
        },
        "address": {
          "type": "string",
          "example": "25 The West Mail Etobicoke, ON M9C 1B8"
        },
        "city": {
          "type": "string",
          "example": "Markham"
        },
        "province": {
          "type": "string",
          "example": "ON"
        },
        "postalCode": {
          "type": "string",
          "example": "L6C 0H9"
        },
        "users": {
          "type": "array",
          "example": [ 3, 4 ]
        }
      },
      "xml": {
        "name": "Location"
      }
    },
    "LocationUpdate": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "1"
        },
        "name": {
          "type": "string",
          "example": "CF Sherway gardens"
        },
        "address": {
          "type": "string",
          "example": "25 The West Mail Etobicoke, ON M9C 1B8"
        },
        "city": {
          "type": "string",
          "example": "Markham"
        },
        "province": {
          "type": "string",
          "example": "ON"
        },
        "postalCode": {
          "type": "string",
          "example": "L6C 0H9"
        },
        "users": {
          "type": "array",
          "example": [ 27, 28 ]
        }
      },
      "xml": {
        "name": "Location"
      }
    },
...
  },
...
}